Remove dead code day calendar and move vm
